[django 1.7] upgraded nodeshot.community.notifications #216
